Remove obsolete tests for callbacks and consent streams
