+++ suites/functional/streams.test.ts
@@ -2,6 +2,7 @@ import {
   verifyAddMemberStream,
   verifyConsentStream,
   verifyConversationStream,
+  verifyGroupConsentStream,
   verifyMembershipStream,
   verifyMessageStream,
   verifyMetadataStream,
@@ -18,91 +19,81 @@ describe(testName, async () => {
   let group: Group;
   let workers = await getWorkers(5);
 
-  it("membership stream", async () => {
-    // Initialize workers
+  it("membership: verify member addition notifications", async () => {
     group = await workers.createGroupBetweenAll();
-
     const verifyResult = await verifyMembershipStream(
       group,
       workers.getAllButCreator(),
       getInboxIds(1),
     );
-
     expect(verifyResult.allReceived).toBe(true);
   });
 
-  it("consent stream", async () => {
+  it("consent: verify consent state changes for direct messages", async () => {
     const verifyResult = await verifyConsentStream(
       workers.getCreator(),
       workers.getReceiver(),
     );
-
     expect(verifyResult.allReceived).toBe(true);
   });
 
-  it("dm stream", async () => {
-    // Create direct message
+  it("consent: verify consent state changes in groups", async () => {
+    group = await workers.createGroupBetweenAll();
+    const verifyResult = await verifyGroupConsentStream(
+      group,
+      workers.getAllButCreator(),
+    );
+    expect(verifyResult.allReceived).toBe(true);
+  });
+
+  it("messages: verify direct message delivery", async () => {
     const creator = workers.getCreator();
     const receiver = workers.getReceiver();
     const newDm = await creator.client.conversations.newDm(
       receiver.client.inboxId,
     );
-
-    // Verify message delivery
     const verifyResult = await verifyMessageStream(newDm as Dm, [receiver], 10);
-
     expect(verifyResult.allReceived).toBe(true);
   });
 
-  it("add member stream", async () => {
+  it("messages: verify group message delivery", async () => {
+    const newGroup = await workers.createGroupBetweenAll();
+    const verifyResult = await verifyMessageStream(
+      newGroup,
+      workers.getAllButCreator(),
+      10,
+    );
+    expect(verifyResult.allReceived).toBe(true);
+  });
+
+  it("metadata: verify group metadata updates", async () => {
+    group = await workers.createGroupBetweenAll();
+    const verifyResult = await verifyMetadataStream(
+      group,
+      workers.getAllButCreator(),
+    );
+    expect(verifyResult.allReceived).toBe(true);
+  });
+
+  it("conversations: verify new conversation notifications", async () => {
+    const verifyResult = await verifyConversationStream(workers.getCreator(), [
+      workers.getReceiver(),
+    ]);
+    expect(verifyResult.allReceived).toBe(true);
+  });
+
+  it("members: verify member addition to existing group", async () => {
     const creator = workers.getCreator();
     const receiver = workers.getReceiver();
-    // Create group with alice as the creator
     group = (await creator.client.conversations.newGroup([
       receiver.client.inboxId,
     ])) as Group;
-    console.log("Group created", group.id);
-
     const addMembers = getInboxIds(1);
     const verifyResult = await verifyAddMemberStream(
       group,
       [receiver],
       addMembers,
     );
-    expect(verifyResult.allReceived).toBe(true);
-  });
-
-  it("message stream", async () => {
-    const newGroup = await workers.createGroupBetweenAll();
-
-    // Verify message delivery
-    const verifyResult = await verifyMessageStream(
-      newGroup,
-      workers.getAllButCreator(),
-      10,
-    );
-
-    expect(verifyResult.allReceived).toBe(true);
-  });
-
-  it("metadata stream", async () => {
-    // Initialize workers
-    group = await workers.createGroupBetweenAll();
-
-    const verifyResult = await verifyMetadataStream(
-      group,
-      workers.getAllButCreator(),
-    );
-
-    expect(verifyResult.allReceived).toBe(true);
-  });
-
-  it("conversation stream", async () => {
-    // Use the dedicated conversation stream verification helper
-    const verifyResult = await verifyConversationStream(workers.getCreator(), [
-      workers.getReceiver(),
-    ]);
-
     expect(verifyResult.allReceived).toBe(true);
   });
 });
